requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==22.0.0

# Base de datos
mysql-connector-python==9.0.0
redis==4.6.0

# Validación y configuración
pydantic>=2.7.0
pydantic-settings==2.5.2

# Generación de reportes
reportlab==4.2.5
openpyxl==3.1.5
pandas==2.2.0
matplotlib==3.9.0

# Observabilidad
prometheus-client==0.19.0

# Seguridad
flask-limiter==3.5.0
PyJWT>=2.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.14.0
